Consolidate xoa/host.go with xoa/data_source_host.go and fix host data source tests
